Packages: Fix connection tests and allow them to run in CI (#12560)
* Use php-mock instead of wp_mock

* Add phpunit script

* Migrate tests from wp_mock to phpmock

* Make tests run in CI

* Run travis only for packages that have the php test dir
