+++ packages/connection/composer.json
@@ -8,8 +8,8 @@
 		"automattic/jetpack-options": "^1.0"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "7.*.*",
-		"10up/wp_mock": "0.4.2"
+		"phpunit/phpunit": "^5.7 || ^6.5 || ^7.5",
+		"php-mock/php-mock": "^2.1"
 	},
 	"autoload": {
 		"psr-4": {
@@ -17,6 +17,9 @@
 		}
 	},
 	"scripts": {
-		"test": "phpunit"
+		"phpunit": [
+			"@composer install",
+			"./vendor/phpunit/phpunit/phpunit --colors=always"
+		]
 	}
 }
