--- packages/connection/tests/bootstrap.php
@@ -1,5 +1,0 @@
-<?php
-
-require_once __DIR__ . '/../vendor/autoload.php';
-
-WP_Mock::bootstrap();
